dialog editor - finish component controller template split
