+++ app/javascript/dialog-editor/box.html
