+++ app/javascript/dialog-editor/dialog-editor.component.js
@@ -1,0 +1,8 @@
+// Top-level dialog editor component.
+export const DialogEditor = {
+  bindings: {
+    treeOptions: '<',
+  },
+  controller: DialogEditorController,
+  template: require('./dialog-editor.html'),
+};
