+++ app/javascript/dialog-editor/dialog-editor.service.js
