+++ app/javascript/dialog-editor/dialog-validation.service.js
