+++ app/javascript/dialog-editor/field.component.js
@@ -1,0 +1,11 @@
+// behaviour for the fields inside of the dialogs boxes
+export const Field = {
+  bindings: {
+    boxPosition: '<',
+    fieldData: '<',
+    setupModalOptions: '&',
+  },
+  controller: FieldController,
+  controllerAs: 'vm',
+  template: require('./field.html'),
+};
