+++ app/javascript/dialog-editor/field.controller.js
@@ -1,4 +1,4 @@
-class FieldController {
+export class FieldController {
   constructor(DialogEditor) {
     this.DialogEditor = DialogEditor;
   }
@@ -27,15 +27,3 @@ class FieldController {
 }
 
 FieldController.$inject = ['DialogEditor'];
-
-// behaviour for the fields inside of the dialogs boxes
-export const Field = {
-  bindings: {
-    boxPosition: '<',
-    fieldData: '<',
-    setupModalOptions: '&',
-  },
-  controller: FieldController,
-  controllerAs: 'vm',
-  template: require('./field.html'),
-};
