+++ app/javascript/dialog-editor/modal-field-template.controller.js
@@ -1,4 +1,4 @@
-class ModalFieldController {
+export class ModalFieldController {
   constructor($element) {
     this.$element = $element;
 
@@ -28,24 +28,3 @@ class ModalFieldController {
 }
 
 ModalFieldController.$inject = ['$element'];
-
-// renders modal templates for each field type
-export const ModalFieldTemplate = {
-  bindings: {
-    addEntry: '=?',
-    categories: '=?',
-    currentCategoryEntries: '=?',
-    modalData: '=',
-    modalTab: '=',
-    modalTabIsSet: '<',
-    removeEntry: '=?',
-    resolveCategories: '=?',
-    setupCategoryOptions: '=?',
-    showFullyQualifiedName: '<',
-    treeOptions: '<',
-  },
-  controller: ModalFieldController,
-  controllerAs: 'vm',
-  scope: true,
-  template: ['$attrs', ($attrs) => require(`./modal-field-template/${$attrs.template}`)],
-};
