+++ app/javascript/dialog-editor/modal.component.js
@@ -1,0 +1,11 @@
+// behaviour for the boxes inside of the dialogs tabs.
+export const Modal = {
+  bindings: {
+    elementInfo: '<',
+    modalOptions: '<',
+    treeOptions: '<',
+  },
+  controller: ModalController,
+  template: '',
+  transclude: true,
+};
