+++ app/javascript/dialog-editor/toolbox.component.js
@@ -1,0 +1,6 @@
+// used as a toolbox for the Dialog Editor
+export const Toolbox = {
+  controller: ToolboxController,
+  controllerAs: 'vm',
+  template: require('./toolbox.html'),
+};
