+++ app/javascript/dialog-editor/toolbox.html
