+++ app/javascript/dialog-editor/validation.component.js
